Add hint in README to use Map object for JWKS cache (#24)
As a first-time user of the library, and the one who is not very fluent in javascript, I was about to write my own implementation for essentially javascript's new Map() object, which can be easily plugged in into jwksCache paramater. 

I think this small description can save some not-fluent-in-js devs a couple or dozen of minutes when trying to make jwks cache work in this library. This PR is more like a feedback from me, decided to share. 

P. S. Another piece of feedback is that I use this library in Lambda@Edge in front of AWS Cloudfront, since the lambda there should be really really small (1MB zipped including node_modules). Thus I couldn't afford to use classic `jsonwebtoken` lib etc.